+++ src/pages/SigninPage/__tests__/SigninPage.test.tsx
@@ -1,21 +1,10 @@
-import { describe, expect, it, vi } from 'vitest';
-import userEvent from '@testing-library/user-event';
+import { describe, expect, it } from 'vitest';
 
 import { render, screen } from 'test/test-utils';
 
 import SigninPage from '../SigninPage';
 
-// mock select functions from react-router-dom
-const mockNavigate = vi.fn();
-vi.mock('react-router-dom', async () => {
-  const original = await vi.importActual('react-router-dom');
-  return {
-    ...original,
-    useNavigate: () => mockNavigate,
-  };
-});
-
-describe('SignIn', () => {
+describe('SigninPage', () => {
   it('should render successfully', async () => {
     // ARRANGE
     render(<SigninPage />);
@@ -23,19 +12,5 @@ describe('SignIn', () => {
 
     // ASSERT
     expect(screen.getByTestId('page-signin')).toBeDefined();
-  });
-
-  it('should navigate upon successful signin', async () => {
-    // ARRANGE
-    render(<SigninPage />);
-    await screen.findByTestId('page-signin');
-
-    // ACT
-    await userEvent.type(screen.getByLabelText('Username'), 'Bret');
-    await userEvent.type(screen.getByLabelText('Password'), 'aB1!12345678');
-    await userEvent.click(screen.getByTestId('signin-form-button-submit'));
-
-    // ASSERT
-    expect(mockNavigate).toHaveBeenCalledWith('/');
   });
 });
